+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v3.6.0 - unreleased
 
+- Deprecated `default_rng` cargo feature and related APIs, most notably, the
+  `new()` constructor and default type parameter `R` of `Scru128Generator`.
 - Deprecated `rand09` cargo feature and the integration of `rand` crate v0.9.
 - Deprecated `Scru128Id::from_fields()` constructor in favor of
   `try_from_fields()`.